requirements.txt
# Core Backend
Django>=4.2,<5.0
fastapi>=0.110.0
uvicorn>=0.29.0

# Environment & Config
python-dotenv>=1.0.1

# Google Drive Integration (OAuth2 + API)
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.125.0

# Document Parsing
PyPDF2>=3.0.0           # PDF parsing
python-docx>=1.1.0      # DOCX parsing

# AI / LLM Integration
openai>=1.30.0

# Data Export
pandas>=2.2.2

# HTTP Client (Django -> FastAPI)
requests>=2.31.0
